// src/types/player.ts
// Represents a player's details
export interface Player {
    plyr: string;         // Player's name
    pys: number;          // Player's score
    w: number;            // Wins
    l: number;            // Losses
    lsOp: string;         // Last opponent
    allOps: string[];     // List of all opponents
  }
  

  